Revert "[chore] split NDMFVRMExporter into several multiple classes (#93)"
This reverts commit d3cb1d18a6c3ed4ff0d3fa04dc3c5b9f81447fc9.
