--- Editor/AssetPathUtils.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7d6fbbb61e384cacb43670e4dca51d9d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
