--- Editor/MaterialBaker.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d5f2ec7d2b06426e999ab2b9ece9eb1c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
