--- Editor/NdmfVrmExporterComponent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a9980d2ec9443cdb2b6b084ea4f7fa0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
