--- Editor/NdmfVrmExporterComponentEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 047def8a89034bd8b814170ed072e8c0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
