--- Editor/NdmfVrmExporterComponentInformation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 87772c58e5a14f2da2699326da6ab6aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
