--- Editor/NdmfVrmExporterPlugin.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c26610e9cd1c4254bfa35baef348a52d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
