--- Editor/Translator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 27035a9533e64de3bafedf6822e7b093
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
