--- Editor/UnityExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e62e28f6153d40a1b39913c6f579a3a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
